+++ lib/hux.dart
@@ -67,6 +67,7 @@ export 'src/widgets/hux_context_menu/hux_context_menu_item.dart';
 export 'src/widgets/hux_context_menu/hux_context_menu_divider.dart';
 export 'src/widgets/hux_date_picker.dart';
 export 'src/widgets/hux_time_picker.dart';
+export 'src/components/dialog/hux_dialog.dart';
 
 // Export external dependencies
 /// FeatherIcons - Beautiful icon set for Flutter applications